Solving import dependency conflict
