+++ freediscovery/datasets.py
@@ -13,6 +13,7 @@ import hashlib
 import platform
 
 import numpy as np
+from .base import PipelineFinder
 
 
 def load_dataset(name='treclegal09_2k_subset', cache_dir='/tmp',
@@ -83,6 +84,8 @@ def load_dataset(name='treclegal09_2k_subset', cache_dir='/tmp',
 
 
     base_url = "http://r0h.eu/d/{}.tar.gz".format(name)
+
+    cache_dir = PipelineFinder._normalize_cachedir(cache_dir)
 
     outdir = os.path.join(cache_dir, name)
     fname = outdir + ".tar.gz"
